+++ src/test/project-test.ts
@@ -1,0 +1,13 @@
+import { describe, it } from "./util/cappucino";
+import { NodeToJSON } from "./util/node2json";
+import { expect } from "chai";
+
+describe("projects", () => {
+    it("work", () => {
+        let file1 = DefaultFile("a");
+        let file2 = DefaultFile("b");
+        expect(false).to.be.true;
+    });
+})
+
+
